pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-crawler"
version = "0.1.0"
description = "网页图片和标题爬虫"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
packages = [{include = "backend"}]

dependencies = [
    # Web框架和API
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    # 认证和安全
    "pyjwt>=2.8.0",
    "werkzeug>=3.0.1",
    # 爬虫相关
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "selenium>=4.15.2",
    "webdriver-manager>=4.0.1",
    # 图片处理
    "pillow>=10.1.0",
    # 数据处理
    "pandas>=2.1.3",
    "numpy>=1.25.2",
    # 配置和工具
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "timeout-decorator>=0.5.0",
    "python-dotenv>=1.0.0",
    # 其他工具
    "colorama>=0.4.6",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.uv]

[tool.hatch.build.targets.wheel]
packages = ["utils", "."]